+++ backend/src/events/routes.ts
@@ -1,10 +1,11 @@
 import { FastifyInstance, RegisterOptions } from 'fastify'
-import { createEvent } from 'controllers/event'
+import { createEvent } from './controller'
+import { DoneFunction } from '../lib/types'
 
 export default function (
   fastify: FastifyInstance,
   opts: RegisterOptions,
-  done: (err?: Error) => void
+  done: DoneFunction
 ) {
   // fastify.get(
   //   '/',
