+++ backend/src/lib/exceptions.ts
@@ -1,15 +1,15 @@
 export class HttpException extends Error {
-  errorCode: number;
+  errorCode: number
 
   constructor(errorCode: number, public readonly message: string) {
-    super(message);
-    this.name = this.constructor.name;
-    this.errorCode = errorCode;
+    super(message)
+    this.name = this.constructor.name
+    this.errorCode = errorCode
   }
 }
 
 export class UniqueConstraintException extends HttpException {
   constructor(public readonly message: string) {
-    super(400, message);
+    super(400, message)
   }
 }
