+++ backend/src/lib/types.ts
@@ -1,5 +1,1 @@
-import { FastifyInstance, RegisterOptions } from 'fastify';
-
-// interface fastifyPlugin {
-//   (fastify: FastifyInstance, opts?: RegisterOptions, done: ((err?:Error)=>void)): void;
-// }
+export type DoneFunction = (err?: Error) => void
