+++ backend/src/organizations/index.ts
@@ -1,0 +1,3 @@
+import routes from './routes'
+
+export default routes
