+++ backend/tsconfig.json
@@ -31,9 +31,9 @@
     "baseUrl": "src/" /* Specify the base directory to resolve non-relative module names. */,
     "paths": {
       // "@/*": ["src/*"],
-      "controllers/*": ["controllers/*"],
-      "models/*": ["models/*"],
-      "routes/*": ["routes/*"]
+      // "controllers/*": ["controllers/*"],
+      // "models/*": ["models/*"],
+      // "routes/*": ["routes/*"]
     } /* Specify a set of entries that re-map imports to additional lookup locations. */,
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
